+++ utils/filesystem/filesystem.h
@@ -6,7 +6,7 @@
 bool fileExists(const char *dataDir, const char *filename);
 unsigned int getFileSize(const char *dataDir, const char *filename);
 bool directoryExists(const char *path);
-unsigned int getDirectorySize(const char *path);
+long getDirectorySize(const char *path);
 void makePath(const char *path);
 
 #endif
